fix(server): extract make_embedder into module to honour 300-line cap
After rebasing on main, src/main.rs grew to 305 lines and tripped the
file-size guard in CI (max 300). Move make_embedder + the fastembed
home_models_dir helper into a sibling embedder_setup.rs and re-export
make_embedder via mod + use. Behaviour unchanged.

main.rs back under the cap; PR #259 CI should now pass.
