+++ crates/convergio-server/src/main.rs
@@ -22,6 +22,9 @@ use std::net::SocketAddr;
 use std::sync::Arc;
 use tracing_subscriber::{fmt, EnvFilter};
 
+mod embedder_setup;
+use embedder_setup::make_embedder;
+
 #[derive(Parser)]
 #[command(name = "convergio", version, about = "Local Convergio daemon", long_about = None)]
 struct Cli {
@@ -205,69 +208,6 @@ fn play_boot_banner() {
     let mut handle = stdout.lock();
     let mut sleeper = boot::RealSleeper;
     let _ = boot::play(&mut handle, &mut sleeper, theme);
-}
-
-/// Construct the daemon's embedder based on `CONVERGIO_EMBED_MODEL`.
-///
-/// - Unset / `deterministic-test` (default): `DeterministicTestEmbedder`
-///   with the canonical 384-dim shape (matches the real-model dim
-///   so a swap doesn't break stored vectors).
-/// - `bge-m3-small-int8` (only when built with `--features fastembed`):
-///   **requested** default from ADR/spec. `fastembed` 5.x does not ship
-///   a 384-dim BGE-M3-small variant, so we currently map this to
-///   `multilingual-e5-small` (384-dim, multilingual) with a warning.
-/// - `multilingual-e5-small` (only when built with `--features fastembed`):
-///   real ONNX via `fastembed-rs`.
-///
-/// FR-3.9 graceful degradation: an unknown value falls back to the test
-/// embedder, never crashes the daemon.
-fn make_embedder() -> Arc<dyn convergio_embed::Embedder> {
-    let model = std::env::var("CONVERGIO_EMBED_MODEL").unwrap_or_default();
-    match model.as_str() {
-        "" | "deterministic-test" => {
-            tracing::info!("embedder: deterministic-test (no model loaded)");
-            Arc::new(convergio_embed::embedder::testing::DeterministicTestEmbedder::new(384))
-        }
-        #[cfg(feature = "fastembed")]
-        requested @ ("bge-m3-small-int8" | "bge-m3-small") => {
-            let cache = home_models_dir();
-            tracing::warn!(
-                requested,
-                "fastembed does not provide a 384-dim bge-m3-small model; using multilingual-e5-small as a compatibility alias"
-            );
-            tracing::info!(?cache, "embedder: multilingual-e5-small (fastembed-rs)");
-            Arc::new(convergio_embed::MultilingualE5Embedder::new(cache))
-        }
-        #[cfg(feature = "fastembed")]
-        "bge-m3" => {
-            let cache = home_models_dir();
-            tracing::info!(?cache, "embedder: bge-m3 (fastembed-rs)");
-            Arc::new(convergio_embed::BgeM3Embedder::new(cache))
-        }
-        #[cfg(feature = "fastembed")]
-        "multilingual-e5-small" => {
-            let cache = home_models_dir();
-            tracing::info!(?cache, "embedder: multilingual-e5-small (fastembed-rs)");
-            Arc::new(convergio_embed::MultilingualE5Embedder::new(cache))
-        }
-        other => {
-            tracing::warn!(
-                model = other,
-                "unknown CONVERGIO_EMBED_MODEL; falling back to deterministic-test"
-            );
-            Arc::new(convergio_embed::embedder::testing::DeterministicTestEmbedder::new(384))
-        }
-    }
-}
-
-#[cfg(feature = "fastembed")]
-fn home_models_dir() -> std::path::PathBuf {
-    let home = std::env::var("HOME").unwrap_or_else(|_| ".".into());
-    let dir = std::path::Path::new(&home).join(".convergio/v3/models");
-    if let Err(e) = std::fs::create_dir_all(&dir) {
-        tracing::warn!(path = %dir.display(), error = %e, "failed to create model cache dir");
-    }
-    dir
 }
 
 fn parse_env_i64(key: &str, default: i64) -> i64 {
